Update README.md
The module version in the example is very old (1.2) but the current module version is 2.2.1.
using terraform 1.4.2 and provider version 4.57.0. this will fail
it's possible to update the module version in the readme to a more recent one to fix this, but it will be another temporary fix which could also be deprecated in the future.
so i'd suggest just removing the version altogether , unless someone would like to use a fixed version.